+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file lists the changes for various versions of the `pprof` gem.
 
+## 1.0.0
+
+* Add support Mac Profiles (`*.provisionprofile`) alongside the already-supported iOS Profiles (`*.mobileprovision`).
+* Add `--platform` filter to only list profiles that support a given platform (`OSX`, `iOS`, …)
+
 ## 0.5.1
 
 * Fix an issue with `-j`/`--json` when used with list mode, so that it can now take `-c` and `-d` flags into account to include (or not) the certificates and devices in the JSON output list mode.